requirements.txt
Django==3.2.0
django-model-utils==4.0.0
